+++ mex/editor/client/src/app/app.routes.ts
@@ -2,6 +2,7 @@ import type { Routes } from "@angular/router";
 
 import { FastTrackActivity } from "./fast-track/fast-track-activity/fast-track-activity";
 import { FastTrackResource } from "./fast-track/fast-track-resource/fast-track-resource";
+import { FigmaTestPage } from "./figma-test-page/figma-test-page";
 import { StartPage } from "./start-page/start-page";
 
 /**
@@ -13,6 +14,8 @@ export const routes: Routes = [
 
   { path: "create/activity", component: FastTrackActivity },
   { path: "create/resource", component: FastTrackResource },
+
+  { path: "figma-test", component: FigmaTestPage },
 
   // Optional: Redirect any unknown URLs back to the root
   { path: "**", redirectTo: "" },