+++ mex/editor/client/src/app/figma-test-page/figma-test-page.ts
@@ -1,0 +1,43 @@
+import { TextFieldModule } from "@angular/cdk/text-field";
+import { Component, signal } from "@angular/core";
+import { FormsModule } from "@angular/forms";
+import { MatButton, MatMiniFabButton } from "@angular/material/button";
+import { MatButtonToggleModule } from "@angular/material/button-toggle";
+import { MatCheckbox } from "@angular/material/checkbox";
+import { MatFormFieldModule } from "@angular/material/form-field";
+import { MatIcon } from "@angular/material/icon";
+import { MatInputModule } from "@angular/material/input";
+import { MatSelectModule } from "@angular/material/select";
+import { MatSlideToggleModule } from "@angular/material/slide-toggle";
+
+@Component({
+  selector: "mex-figma-test-page",
+  imports: [
+    FormsModule,
+    MatButton,
+    MatButtonToggleModule,
+    MatCheckbox,
+    MatFormFieldModule,
+    MatIcon,
+    MatInputModule,
+    MatMiniFabButton,
+    MatSelectModule,
+    MatSlideToggleModule,
+    TextFieldModule,
+  ],
+  templateUrl: "./figma-test-page.html",
+  styleUrl: "./figma-test-page.scss",
+})
+/**
+ * FigmaTestPage for the app.
+ */
+export class FigmaTestPage {
+  description = signal("");
+  value = signal<number | null>(null);
+  options = [
+    { label: "None", value: null },
+    { label: "One", value: 1 },
+    { label: "Two", value: 2 },
+    { label: "Three", value: 3 },
+  ];
+}