+++ mex/editor/client/src/app/nav-bar/nav-bar.ts
@@ -1,21 +1,21 @@
 import { Component } from "@angular/core";
+import { MatIcon } from "@angular/material/icon";
 import { MatSelectModule } from "@angular/material/select";
 import { MatToolbar } from "@angular/material/toolbar";
 import { RouterLink, RouterLinkActive } from "@angular/router";
 import { TranslocoDirective } from "@jsverse/transloco";
 
 import { LanguageSelector } from "./language-selector/language-selector";
-import { ThemeToggle } from "./theme-toggle/theme-toggle";
 
 @Component({
   selector: "mex-nav-bar",
   imports: [
     LanguageSelector,
+    MatIcon,
     MatSelectModule,
     MatToolbar,
     RouterLink,
     RouterLinkActive,
-    ThemeToggle,
     TranslocoDirective,
   ],
   templateUrl: "./nav-bar.html",